Refactoring - moves FakeTokenWithTrivia to separate file
